Commit 019 Nick.vonklemperer@netronome.com Separations for 2 and 3 node cluster

+++ scripts/08_patch_image.sh
@@ -23,7 +23,11 @@ fi
 
 # Download and extract vrouter build files
 echo "Downloading patches"
-wget http://pahome.netronome.com/releases-intern/vrouter/builds/${BUILD_NAME}.tar
+wget --tries=1 http://pahome.netronome.com/releases-intern/vrouter/builds/${BUILD_NAME}.tar
+if [[ $? -ne 0 ]]; then
+    exit 1;
+fi
+
 tar -xvf Netronome_R4.1_build_**.tar
 rm -r Netronome_R4.1_build_**.tar
